RenderMatrix: replace assert raised when building without intrinsics
That assert can be easily reproduced with the map mars_city2. It
raises just when the door opens if you built with USE_INTRINSICS=OFF.

Was able to reproduce it using the Mesa vulkan driers, on the rpi4 (so
USE_INTRINSICS is needed) and with Intel GPU (so USE_INTRINSICS is not
really needed).